+++ packages/hardhat-helpers/src/constants.js
@@ -1,11 +1,6 @@
-const {
-  ADDRESS_ZERO,
-  MAX_UINT,
-  MAX_GAS,
-} = require('@unlock-protocol/hardhat-helpers')
-
 module.exports = {
-  ADDRESS_ZERO,
-  MAX_UINT,
-  MAX_GAS,
+  ADDRESS_ZERO: '0x0000000000000000000000000000000000000000',
+  MAX_UINT:
+    '115792089237316195423570985008687907853269984665640564039457584007913129639935',
+  MAX_GAS: 6700000,
 }
