+++ smart-contracts/test/Unlock/createLockAtVersion.js
@@ -2,7 +2,7 @@ const { expect, assert } = require('chai')
 const { ethers, upgrades } = require('hardhat')
 const { expectRevert } = require('@openzeppelin/test-helpers')
 const createLockHash = require('../helpers/createLockCalldata')
-const { ADDRESS_ZERO } = require('../helpers/constants')
+const { ADDRESS_ZERO } = require('../helpers')
 // lock args
 const args = [
   60 * 60 * 24 * 30, // expirationDuration: 30 days
