+++ smart-contracts/test/fixtures/locks.js
@@ -1,5 +1,5 @@
 const { ethers } = require('hardhat')
-const { MAX_UINT } = require('../helpers/constants')
+const { MAX_UINT } = require('../helpers')
 
 let publicLock = {
   expirationDuration: 60 * 60 * 24 * 30, // 30 days
