+++ smart-contracts/test/helpers/constants.js
@@ -1,0 +1,11 @@
+const {
+  ADDRESS_ZERO,
+  MAX_UINT,
+  MAX_GAS,
+} = require('@unlock-protocol/hardhat-helpers')
+
+module.exports = {
+  ADDRESS_ZERO,
+  MAX_UINT,
+  MAX_GAS,
+}
